adding postprocessing improvements (#437)
* adding postprocessing improvements

* add skellyforge to dependancies

* Update process_recording_folder.py

* disabling SkellyForge rotation

* fixing conflicts

* deleting unnecessary files

* adding body segments to data loader

* create 'data_models' folder and move stuff into it

Arguably this could be moved the 'data_layer' folder
, but I think it makes sense to keep it close to the classes/functions that will use it (esp because this will likely move to `skellytracker` soon enough)

I

* importing mediapipe marker names from `mediapipe_stuff` folder and adding a check to make sure the mediapipe folks don't change the names out from under us

---------
