+++ freemocap/core_processes/process_motion_capture_videos/process_recording_folder.py
@@ -17,26 +17,23 @@ from freemocap.core_processes.detecting_things_in_2d_images.mediapipe_stuff.conv
 from freemocap.core_processes.detecting_things_in_2d_images.mediapipe_stuff.mediapipe_skeleton_detector import (
     MediaPipeSkeletonDetector,
 )
-from freemocap.core_processes.detecting_things_in_2d_images.mediapipe_stuff.mediapipe_skeleton_names_and_connections import (
+from freemocap.core_processes.detecting_things_in_2d_images.mediapipe_stuff.data_models.mediapipe_skeleton_names_and_connections import (
     mediapipe_names_and_connections_dict,
 )
+from freemocap.core_processes.post_process_skeleton_data.calculate_center_of_mass import run_center_of_mass_calculations
 from freemocap.core_processes.post_process_skeleton_data.estimate_skeleton_segment_lengths import (
     estimate_skeleton_segment_lengths,
     mediapipe_skeleton_segment_definitions,
 )
-from freemocap.core_processes.post_process_skeleton_data.gap_fill_filter_and_origin_align_skeleton_data import (
-    gap_fill_filter_origin_align_3d_data_and_then_calculate_center_of_mass,
-)
+from freemocap.core_processes.post_process_skeleton_data.post_process_skeleton import post_process_data, \
+    save_skeleton_array_to_npy
 from freemocap.core_processes.post_process_skeleton_data.process_single_camera_skeleton_data import \
     process_single_camera_skeleton_data
 from freemocap.data_layer.data_saver.data_saver import DataSaver
-
 from freemocap.data_layer.recording_models.post_processing_parameter_models import PostProcessingParameterModel
-from freemocap.system.paths_and_filenames.file_and_folder_names import (
-    MEDIAPIPE_BODY_3D_DATAFRAME_CSV_FILE_NAME,
-    RAW_DATA_FOLDER_NAME,
-)
-
+from freemocap.system.paths_and_filenames.file_and_folder_names import RAW_DATA_FOLDER_NAME, \
+    SEGMENT_CENTER_OF_MASS_NPY_FILE_NAME, CENTER_OF_MASS_FOLDER_NAME, TOTAL_BODY_CENTER_OF_MASS_NPY_FILE_NAME, \
+    MEDIAPIPE_BODY_3D_DATAFRAME_CSV_FILE_NAME
 from freemocap.tests.test_image_tracking_data_shape import (
     test_image_tracking_data_shape,
 )
@@ -71,7 +68,7 @@ def process_recording_folder(
 
     if rec.mediapipe_parameters_model.skip_2d_image_tracking:
         logger.info(f"Skipping 2d skeleton detection and loading data from: {rec.recording_info_model.mediapipe_2d_data_npy_file_path}")
-        try: 
+        try:
             mediapipe_image_data_numCams_numFrames_numTrackedPts_XYZ = np.load(rec.recording_info_model.mediapipe_2d_data_npy_file_path)
         except Exception as e:
             logger.error("Failed to load 2D data, cannot continue processing")
@@ -99,19 +96,21 @@ def process_recording_folder(
             image_tracking_data_file_path=rec.recording_info_model.mediapipe_2d_data_npy_file_path,
         )
     except AssertionError as error_message:
-            logger.error(error_message)
+        logger.error(error_message)
 
-
+    # spoof 3D data if single camera
     if mediapipe_image_data_numCams_numFrames_numTrackedPts_XYZ.shape[0] == 1:
         # spoof 3D data if single camera
         (raw_skel3d_frame_marker_xyz, skeleton_reprojection_error_fr_mar) = process_single_camera_skeleton_data(
             input_image_data_frame_marker_xyz=mediapipe_image_data_numCams_numFrames_numTrackedPts_XYZ[0],
             raw_data_folder_path=Path(rec.recording_info_model.raw_data_folder_path))
+
     else:
         if rec.anipose_triangulate_3d_parameters_model.skip_3d_triangulation:
             logger.info(f"Skipping 3d triangulation and loading data from: {rec.recording_info_model.raw_mediapipe_3d_data_npy_file_path}")
             raw_skel3d_frame_marker_xyz = np.load(rec.recording_info_model.raw_mediapipe_3d_data_npy_file_path)
-            skeleton_reprojection_error_fr_mar = np.load(rec.recording_info_model.mediapipe_reprojection_error_data_npy_file_path)
+            skeleton_reprojection_error_fr_mar = np.load(
+                rec.recording_info_model.mediapipe_reprojection_error_data_npy_file_path)
         else:
             logger.info("Triangulating 3d skeletons...")
 
@@ -143,23 +142,33 @@ def process_recording_folder(
     except AssertionError as error_message:
         logger.error(error_message)
 
+    logger.info("Using SkellyForge Post-Processing to clean up data...")
 
-    #rotate so skeleton is closer to 'vertical' in a z-up reference frame
-    rotated_raw_skel3d_frame_marker_xyz = rotate_by_90_degrees_around_x_axis(raw_skel3d_frame_marker_xyz) 
+    rotated_raw_skel3d_frame_marker_xyz = rotate_by_90_degrees_around_x_axis(raw_skel3d_frame_marker_xyz)
+
+    skel3d_frame_marker_xyz = post_process_data(
+        recording_processing_parameter_model=recording_processing_parameter_model,
+        raw_skel3d_frame_marker_xyz= rotated_raw_skel3d_frame_marker_xyz)
+
+    path_to_folder_where_we_will_save_this_data = rec.recording_info_model.output_data_folder_path
 
 
-    logger.info("Gap-filling, butterworth filtering, origin aligning 3d skeletons, then calculating center of mass ...")
 
-    skel3d_frame_marker_xyz = gap_fill_filter_origin_align_3d_data_and_then_calculate_center_of_mass(
-        raw_skel3d_frame_marker_xyz=rotated_raw_skel3d_frame_marker_xyz,
-        skeleton_reprojection_error_fr_mar=skeleton_reprojection_error_fr_mar,
-        path_to_folder_where_we_will_save_this_data=rec.recording_info_model.output_data_folder_path,
-        skip_butterworth_filter=rec.post_processing_parameters_model.skip_butterworth_filter,
-        sampling_rate=rec.post_processing_parameters_model.framerate,
-        cut_off=rec.post_processing_parameters_model.butterworth_filter_parameters.cutoff_frequency,
-        order=rec.post_processing_parameters_model.butterworth_filter_parameters.order,
-        reference_frame_number=None,
-    )
+
+    logger.info("Saving post-processed data")
+    save_skeleton_array_to_npy(array_to_save=skel3d_frame_marker_xyz, skeleton_file_name="mediaPipeSkel_3d_body_hands_face.npy",
+                               path_to_folder_where_we_will_save_this_data=path_to_folder_where_we_will_save_this_data)
+
+    segment_COM_frame_imgPoint_XYZ, totalBodyCOM_frame_XYZ = run_center_of_mass_calculations(processed_skel3d_frame_marker_xyz=skel3d_frame_marker_xyz)
+
+    logger.info("Saving segment center of mass data")
+    save_skeleton_array_to_npy(array_to_save=segment_COM_frame_imgPoint_XYZ, skeleton_file_name=SEGMENT_CENTER_OF_MASS_NPY_FILE_NAME,
+                               path_to_folder_where_we_will_save_this_data= Path(path_to_folder_where_we_will_save_this_data)/CENTER_OF_MASS_FOLDER_NAME)
+
+    logger.info("Saving total body center of mass data")
+    save_skeleton_array_to_npy(array_to_save=totalBodyCOM_frame_XYZ, skeleton_file_name=TOTAL_BODY_CENTER_OF_MASS_NPY_FILE_NAME,
+                               path_to_folder_where_we_will_save_this_data= Path(path_to_folder_where_we_will_save_this_data)/CENTER_OF_MASS_FOLDER_NAME)
+
 
     try:
         test_mediapipe_skeleton_data_shape(
@@ -169,7 +178,6 @@ def process_recording_folder(
         )
     except AssertionError as error_message:
         logger.error(error_message)
-
 
     logger.info("Breaking up big `npy` into smaller bits and converting to `csv`...")
     # break up big NPY and save out csv's
